+++ NEWS.md
@@ -7,6 +7,7 @@
 - Thanks to @tjmckinley, an issue relating to `libstdc++.so.6` on Linux machines when R is restarting after the installation of the package is fixed.
 - `alm()` and `mice()` can locate new design points for stochastic simulators with (D)GP or bundle emulators that can deal with stochastic outputs.
 - `design()` can be used to construct (D)GP or bundle emulators adaptively by utilizing multiple realizations from a stochastic simulator at the same design positions through the new argument `reps` when `method = alm` or `method = mice`.
+- A new slot called `specs` is added to the objects returned by `gp()` and `dgp()` that contains the key information of the kernel functions used in the constructions of GP and DGP emulators.
 
 # dgpsi 2.2.0